refactor(web): modularize app and paginate device loading
- split monolithic web.py into webapp server/templates modules

- keep web.py as compatibility entrypoint exporting WebServer

- add paginated, filtered, sorted device queries in db layer

- update dashboard API and UI to use server-side pagination
