+++ bluehood/webapp/__init__.py
@@ -1,0 +1,5 @@
+"""Bluehood web app package."""
+
+from .server import WebServer
+
+__all__ = ["WebServer"]
